refactor: reorganized the distribution_utils submodule
