--- src/hssm/distribution_utils/onnx/__init__.py
@@ -1,19 +1,0 @@
-"""Utility functions for creating pytensor Ops from onnx model files."""
-
-from .onnx import (
-    make_jax_logp_funcs_from_jax_callable,
-    make_jax_logp_funcs_from_onnx,
-    make_jax_logp_ops,
-    make_pytensor_logp,
-)
-from .onnx2pt import pt_interpret_onnx
-from .onnx2xla import interpret_onnx
-
-__all__ = [
-    "interpret_onnx",
-    "make_jax_logp_funcs_from_onnx",
-    "make_jax_logp_funcs_from_jax_callable",
-    "make_jax_logp_ops",
-    "make_pytensor_logp",
-    "pt_interpret_onnx",
-]
